Fix C4244/C4267: add explicit casts for narrowing conversions on x64
All changes add an explicit (int), (DWORD), or (unsigned int) cast to make
intentional narrowing visible to the compiler.  No logic is changed.

C4244 - conversion with possible loss of data:
  IOGrunt.cpp:1033      __time64_t -> int  (elapsed ms in transfer_delay)
  IOManager.cpp:1078    DWORDLONG  -> unsigned int  (srand seed)
  IOPortTCP.cpp:916     SOCKET     -> int  (select nfds arg; ignored on Windows)
  AccessSpecList.cpp    INT_PTR    -> int  (GetSize() return / local vars)
  Manager.cpp           INT_PTR    -> int  (GetSize() in WorkerCount/InterfaceCount)
  ManagerList.cpp:982   INT_PTR    -> int  (GetSize() in ManagerCount)
  ManagerMap.cpp        INT_PTR    -> int  (GetSize() in Store/Retrieve/etc.)
  Worker.cpp            INT_PTR    -> int  (GetSize() in TargetCount/AccessSpecCount)

C4267 - size_t -> smaller integer:
  IOManagerWin.cpp:674  size_t -> DWORD  (strlen -> DWORD length)
  IOManagerWin.cpp:1106 size_t -> DWORD  (strlen -> DWORD str_length)
  IOManagerWin.cpp:1150 size_t -> int    (strlen -> int loop counter)
  IOManagerWin.cpp:1299 size_t -> int    (strlen -> int str_length)

Build now produces 0 warnings, 0 errors.

+++ src/AccessSpecList.cpp
@@ -768,7 +768,7 @@ Test_Spec *AccessSpecList::Get(int index)
 //
 int AccessSpecList::Count()
 {
-	return spec_list.GetSize();
+	return (int)spec_list.GetSize();
 }
 
 //
@@ -777,7 +777,7 @@ int AccessSpecList::Count()
 Test_Spec *AccessSpecList::RefByName(const char *check_name)
 {
 	Test_Spec *spec;
-	int spec_count = spec_list.GetSize();
+	int spec_count = (int)spec_list.GetSize();
 
 	for (int s = 0; s < spec_count; s++) {
 		spec = Get(s);
@@ -795,7 +795,7 @@ Test_Spec *AccessSpecList::RefByName(const char *check_name)
 //
 int AccessSpecList::IndexByRef(const Test_Spec * spec)
 {
-	int spec_count = spec_list.GetSize();
+	int spec_count = (int)spec_list.GetSize();
 
 	for (int index = 0; index < spec_count; index++) {
 		if (Get(index) == spec)
